tests/wpt/css-tests/css-transforms-1_dev/html/transform_translate_second_omited.htm

<!DOCTYPE html>
<html><head>
    <title>CSS Transforms API Test: transform translate with second translation value omited</title>
    <link href="http://www.intel.com" rel="author" title="Intel">
    <link href="http://www.w3.org/TR/css-transforms-1/#transform-property" rel="help">
    <link href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions" rel="help">
    <meta content="dom" name="flags">
    <meta content="Check if transform supports translate(100px)" name="assert"> 
    <script src="/resources/testharness.js" type="text/javascript"></script>
    <script src="/resources/testharnessreport.js" type="text/javascript"></script>    
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
    test(function() {
        document.getElementById("test").style.transform = "translate(100px)";
        var value = document.getElementById("test").style.getPropertyCSSValue("transform").cssText;
        assert_equals(value, "translate(100px)")
    }, "transform_translate_100px");
    </script>    


</body></html>
